fix more warnings

+++ brn2.c
@@ -296,7 +296,7 @@ brn2_list_from_file(FileList *list, char *filename, bool is_old) {
                   lines_stat.st_size);
             fatal(EXIT_FAILURE);
         }
-        if (lines_stat.st_size >= UINT32_MAX) {
+        if (lines_stat.st_size >= (int64)UINT32_MAX) {
             error("Error: File size = %ld.\n", lines_stat.st_size);
             fatal(EXIT_FAILURE);
         }
